+++ x/auth/internal/types/codec.go
@@ -13,7 +13,6 @@ func RegisterCodec(cdc *codec.Codec) {
 
 	// Register custom types
 	cdc.RegisterConcrete(MsgManualVesting{}, "auth/MsgManualVesting", nil)
-	cdc.RegisterConcrete(MsgSendLock{}, "auth/MsgSendLock", nil)
 }
 
 func RegisterAccountTypeCodec(o interface{}, name string) {
