+++ x/auth/internal/types/msgs.go
@@ -3,11 +3,12 @@ package types
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	"github.com/cosmos/cosmos-sdk/x/auth"
 )
 
 const (
-	ModuleName = "auth"
-	RouterKey  = ModuleName
+	ModuleName = auth.ModuleName
+	RouterKey = ModuleName
 )
 
 
@@ -54,50 +55,4 @@ func (m MsgManualVesting) GetSignBytes() []byte {
 // GetSigners defines whose signature is required.
 func (m MsgManualVesting) GetSigners() []sdk.AccAddress {
 	return []sdk.AccAddress{m.Certifier}
-}
-
-// MsgSendLock transfers coins and have them vesting
-// in the receiver's manual vesting account.
-type MsgSendLock struct {
-	From   sdk.AccAddress `json:"from" yaml:"from"`
-	To     sdk.AccAddress `json:"to" yaml:"to"`
-	Amount sdk.Coin       `json:"amount" yaml:"amount"`
-}
-
-var _ sdk.Msg = MsgSendLock{}
-
-// NewMsgSendLock returns a MsgSendLock object.
-func NewMsgSendLock(from, to sdk.AccAddress, amount sdk.Coin) MsgSendLock {
-	return MsgSendLock{
-		From:   from,
-		To:     to,
-		Amount: amount,
-	}
-}
-
-// Route returns the name of the module.
-func (m MsgSendLock) Route() string { return ModuleName }
-
-// Type returns a human-readable string for the message.
-func (m MsgSendLock) Type() string { return "send_locked" }
-
-// ValidateBasic runs stateless checks on the message.
-func (m MsgSendLock) ValidateBasic() error {
-	if m.From.Empty() {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, "missing from address")
-	}
-	if m.To.Empty() {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidAddress, "missing to address")
-	}
-	return nil
-}
-
-// GetSignBytes encodes the message for signing.
-func (m MsgSendLock) GetSignBytes() []byte {
-	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(m))
-}
-
-// GetSigners defines whose signature is required.
-func (m MsgSendLock) GetSigners() []sdk.AccAddress {
-	return []sdk.AccAddress{m.From}
 }
